Fixed display bug in progress display

+++ src/threading/Worker.cpp
@@ -38,7 +38,7 @@ namespace threading {
 		if (Application::getInstance().getVerbosity() > boost::log::trivial::info) {
 			char buffer[80];
 			sprintf(buffer, "Progress: %d/%d (%4.2f\%)", r.rayNumber, Application::getInstance().numWorkers,
-					100*r.rayNumber/Application::getInstance().numWorkers);
+					100.0*r.rayNumber/((double)Application::getInstance().numWorkers));
 			CommandLine::getInstance().updateBody(buffer);
 		}
 	}
